Fix walberla code and remove warnings on C++ code

+++ time.h
@@ -19,14 +19,14 @@ auto calculate_time_difference(std::chrono::high_resolution_clock::time_point be
 
 double compute_mean(std::vector<double> const& v) {
     double sum = std::accumulate(v.begin(), v.end(), 0.0);
-    return sum / v.size();
+    return sum / (double) v.size();
 }
 
 double compute_standard_deviation(std::vector<double> const& v, double const mean) {
     std::vector<double> diff(v.size());
     std::transform(v.begin(), v.end(), diff.begin(), [mean](double x) { return x - mean; });
     double sq_sum = std::inner_product(diff.begin(), diff.end(), diff.begin(), 0.0);
-    return std::sqrt(sq_sum / v.size());
+    return std::sqrt(sq_sum / (double) v.size());
 }
